refactor: move GPU compute files into subdirectory
