+++ src/compute/gpu/gpu_platform.cpp
