+++ src/compute/gpu/gpu_platform.h
