+++ src/compute/gpu/gpu_transparency.cpp
